chore: updated build script to generate bundles (#1902)

+++ scripts/upload-apk.sh
@@ -8,59 +8,57 @@ export DEVELOPMENT_BRANCH=${DEVELOPMENT_BRANCH:-development}
 git config --global user.email "noreply@travis.com"
 git config --global user.name "Travis CI" 
 
+if [ "$TRAVIS_PULL_REQUEST" != "false" -o "$TRAVIS_REPO_SLUG" != "fossasia/pslab-android" ] || ! [ "$TRAVIS_BRANCH" == "$DEVELOPMENT_BRANCH" -o "$TRAVIS_BRANCH" == "$PUBLISH_BRANCH" ]; then
+    echo "We upload apk only for changes in Development or Master"
+    exit 0
+fi
+
+# Generate Playstore bundle
+./gradlew bundlePlaystoreRelease
+
 # #clone the repository
 git clone --quiet --branch=apk https://fossasia:$GITHUB_API_KEY@github.com/fossasia/pslab-android apk > /dev/null
 cd apk
 
 if [ "$TRAVIS_BRANCH" == "$PUBLISH_BRANCH" ]; then
-    rm -rf *
-else 
+    rm -rf pslab-master*
+else
     rm -rf pslab-dev*
 fi
 
-# Signing Apps
+find ../app/build/outputs -type f -name '*.apk' -exec cp -v {} . \;
+find ../app/build/outputs -type f -name '*.aab' -exec cp -v {} . \;
 
-if [ "$TRAVIS_BRANCH" == "$PUBLISH_BRANCH" ]; then
-    ./gradlew bundlePlaystoreRelease
-
-    find ../app/build/outputs -type f \( -name "*.aab" -o -name ".apk" \) -exec cp {} . \;
-
-    echo "Push to master branch detected, signing the app..."
-    # Retain apk files for testing
-    mv app-playstore-debug.apk pslab-master-debug.apk
-    mv app-fdroid-debug.apk pslab-master-debug-fdroid.apk
-    # Generate temporary apk for signing
-    mv app-playstore-release.apk pslab-master-release.apk
-    mv app-fdroid-release.apk pslab-master-release-fdroid.apk
-    mv app.aab pslab-master.aab
-fi
-
-if [ "$TRAVIS_BRANCH" == "$DEVELOPMENT_BRANCH" ]; then
-    find ../app/build/outputs -type f -name "*.apk" -exec cp {} . \;
-
-    echo "Push to development branch detected, generating apk..."
-    # Rename apks with dev prefixes
-    mv app-playstore-debug.apk pslab-dev-debug.apk
-    mv app-fdroid-debug.apk pslab-dev-debug-fdroid.apk
-
-    mv app-playstore-release.apk pslab-dev-release.apk
-    mv app-fdroid-release.apk pslab-dev-release-fdroid.apk
-fi
+for file in app*; do
+    if [ "$TRAVIS_BRANCH" == "$PUBLISH_BRANCH" ]; then
+        if [[ ${file} =~ ".aab" ]]; then
+            mv $file pslab-master-${file}
+        else
+            mv $file pslab-master-${file:4}
+        fi
+    elif [ "$TRAVIS_BRANCH" == "$DEVELOPMENT_BRANCH" ]; then
+        if [[ ${file} =~ ".aab" ]]; then
+                mv $file pslab-dev-${file}
+        else
+                mv $file pslab-dev-${file:4}
+        fi
+    fi
+done
 
 git checkout --orphan temporary
 # Push generated apk files to apk branch
 git add .
-git commit -am "Travis build pushed [$TRAVIS_BRANCH]"
+git commit -m "Travis build pushed to [$TRAVIS_BRANCH]"
 
 # Delete current apk branch
 git branch -D apk
 # Rename current branch to apk
 git branch -m apk
 
-git push origin apk --force --quiet > /dev/null
+git push origin apk -f --quiet > /dev/null
 
 # Publish App to Play Store
 if [ "$TRAVIS_BRANCH" == "$PUBLISH_BRANCH" ]; then
     gem install fastlane
-    fastlane supply --aab pslab-master.aab --track alpha --json_key ../scripts/fastlane.json --package_name $PACKAGE_NAME
+    fastlane supply --aab pslab-master-app.aab --track alpha --json_key ../scripts/fastlane.json --package_name $PACKAGE_NAME
 fi
